Bug fixes/enhancements to animations/behaviors
In Landscape Item behaviors, this.position works.
In animations, there's a memory object that persists across time and
can be initialized on the first tick.
In animations, "view.func()" can now just be "func()" due to a "with"
wrapper.
Animations with errors just dump them once to the client console, then
stop running.