Tutorial Areas and more (see comments)
Tutorial areas in game database not part of this commit, because the
game database is not in the repository.
But to go along with those tutorial areas, maker commands now link to
an extensive website of maker reference documents and tutorials, which
are in this commit, and in the repository.