Adding "no CDNs" option to settings
This makes it so that you can store local copies of scripts and
stylesheets that you would otherwise get from CDNs, and if you turn
this option on, it will use your server's copies rather than fetching
them from the web.  This is good for playing a local game without an
Internet connection, or working on your game world on an airplane.